backend/requirements.txt
fastapi
uvicorn
requests
python-multipart
python-dotenv
openai
beautifulsoup4
lxml
pandas
tiktoken